+++ src/components/checkbox/interfaces/index.ts
@@ -1,4 +1,5 @@
 export type {
+  CheckboxProps,
   CheckboxRootProps,
   CheckboxIndicatorProps,
   CheckboxLabelProps,