+++ src/components/radio/index.ts
@@ -1,8 +1,10 @@
 export { Radio } from './core/radio';
 export type {
+  RadioProps,
   RadioRootProps,
   RadioIndicatorProps,
   RadioLabelProps,
   RadioDescriptionProps,
   RadioSize,
+  RadioVariant,
 } from './interfaces/RadioProps';